requirements.txt
pydantic>=2.0.0
google-generativeai
